// src/app/Models/extraapplicantinfo.ts
export class Extraapplicantinfo {
    ID: number;
    CLIENT_ID: number;
    PROPOSAL_ID: number;
    EXTRA_INFORMATION_ID: number;
    IS_PROVIDED: any;
    IS_PROVIDED_NULL: any;
    PROVIDED_DATETIME: Date;
    IS_COMPLUSORY: boolean;
    IS_VERIFIED: any;
    IS_APPROVED: any;
    REMARK: string;
    USER_ID: number;
    CURRENT_STAGE_ID: number
}